requirements.txt
discord.py>=2.5.2
pynacl==1.5.0
python-dotenv>=0.9.9
pytchat>=0.5.5
httpx==0.28.1
gTTS==2.5.4
langid==1.1.6
pydub==0.25.1
